git-svn-id: svn://svn.dd-wrt.com/DD-WRT@49669 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/mac80211/net/wireless/rdev-ops.h
@@ -330,7 +330,7 @@ rdev_get_mesh_config(struct cfg80211_registered_device *rdev,
 
 static inline int
 rdev_update_mesh_config(struct cfg80211_registered_device *rdev,
-			struct net_device *dev, u64 mask,
+			struct net_device *dev, u32 mask,
 			const struct mesh_config *nconf)
 {
 	int ret;
